refine code

+++ sea-core-basic/src/main/java/com/github/seaxlab/core/util/IntegerUtil.java
@@ -80,7 +80,7 @@ public final class IntegerUtil {
     if (value == null) {
       return false;
     }
-    return value.intValue() != 0L;
+    return value != 0L;
   }
 
   /**
